refactor: complete SQL-first architecture restructuring
- Add complete sql/ directory with DDL, ETL, and query templates
- Add FastAPI application structure in api/ directory
- Add Airflow DAG for SQL-based ETL pipeline
- Reorganize src/ticker_converter/ modules (etl, database, api_clients)
- Remove storage system and obsolete complexity
- Update test structure to match new architecture
- Fix all import paths and dependencies

Closes #12
