+++ api/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI application package for NYSE stock analytics."""
