+++ copilot-instructions.md
@@ -142,4 +142,5 @@ The project uses a **Feature Branch Workflow** for development and relies heavil
 *   **Be specific**: When generating code, be as specific as possible about requirements and provide examples if needed [26, 36].
 *   **Validate suggestions**: Always assume a human will review and validate the generated code for functionality, security, readability, and maintainability [37].
 *   **Keep responses concise**: While detailed, aim for short, self-contained statements that are broadly applicable. Avoid overly long responses or internal loops, as very long instruction files (e.g., 9,000 tokens) can lead to slower responses and unexpected behavior [1, 38, 39].
+*   **NO EMOJIS**: Never use emojis in any responses, comments, commit messages, documentation, or code suggestions. Maintain professional, text-only communication throughout all interactions.
 *   **Modular Instructions (VS Code Specific)**: If your IDE supports it (e.g., VS Code), you can consider using multiple `.instructions.md` files within a `.github/instructions` directory, with `applyTo` frontmatter, to provide context relevant to specific files or directories [18-20, 40]. This can help manage context for very large codebases [39].
