+++ pyproject.toml
@@ -18,15 +18,13 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Topic :: Office/Business :: Financial",
     "Topic :: Scientific/Engineering :: Information Analysis",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 dependencies = [
-    # Core Data Processing (minimal for API work)
+    # Core Data Processing (minimal for SQL-first approach)
     "pandas>=2.0.0",
     "numpy>=1.24.0",
     # API Clients for Financial Data
@@ -34,9 +32,9 @@ dependencies = [
     "python-dotenv>=1.0.0",
     # Data Models and Validation
     "pydantic>=2.0.0",
-    # Storage (for Issue #2)
+    # Storage for raw data
     "pyarrow>=12.0.0",
-    # Testing Framework (basic)
+    # Testing Framework
     "pytest>=7.4.0",
     # Code Quality and Linting
     "black>=23.7.0",
@@ -51,25 +49,13 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-# Additional packages for later GitHub issues
-storage = [
-    "pyarrow>=12.0.0",  # For Parquet support (Issue #2)
-]
+# SQL-first architecture dependencies
 database = [
-    "psycopg2-binary>=2.9.0",  # PostgreSQL adapter (Issue #4)
-    "sqlalchemy>=2.0.0",      # ORM and database toolkit
-    "alembic>=1.11.0",        # Database migrations
-]
-airflow = [
-    "apache-airflow>=2.7.0",                      # (Issue #5)
-    "apache-airflow-providers-postgres>=5.6.0",
-    "apache-airflow-providers-http>=4.5.0",
-]
-models = [
-    "pydantic>=2.0.0",  # Data validation (Issue #3)
+    "psycopg2-binary>=2.9.0",  # PostgreSQL adapter
+    "asyncpg>=0.28.0",         # Async PostgreSQL adapter
 ]
 api = [
-    "fastapi>=0.100.0",        # Web Framework (Issue #6)
+    "fastapi>=0.100.0",        # Web Framework
     "uvicorn[standard]>=0.23.0",
     "httpx>=0.24.0",           # For testing FastAPI
 ]
@@ -79,51 +65,10 @@ dev = [
     "pytest-cov>=4.1.0",
     "pytest-mock>=3.11.0",
     "responses>=0.24.0",
-    "faker>=19.0.0",
-    # Development tools
-    "ipython>=8.14.0",
-    "jupyter>=1.0.0",
-    "jupyterlab>=4.0.0",
-    "nbconvert>=7.7.0",
-    # Code quality & security
-    "pre-commit>=3.3.0",
-    "bandit>=1.7.5",
-    "safety>=2.3.0",
-    # Documentation
-    "sphinx>=7.1.0",
-    "sphinx-rtd-theme>=1.3.0",
-    "sphinx-autodoc-typehints>=1.24.0",
-    # Performance profiling
-    "memory-profiler>=0.61.0",
-    "line-profiler>=4.0.0",
-    # Database tools
-    "pgcli>=3.5.0",
-]
-cloud = [
-    "boto3>=1.28.0",
-    "google-cloud-storage>=2.10.0",
-]
-viz = [
-    "matplotlib>=3.7.0",
-    "seaborn>=0.12.0",
-    "plotly>=5.15.0",
-]
-auth = [
-    "python-jose[cryptography]>=3.3.0",
-    "passlib[bcrypt]>=1.7.4",
-    "python-multipart>=0.0.6",
-]
-monitoring = [
-    "structlog>=23.1.0",
-    "prometheus-client>=0.17.0",
-]
-async = [
-    "aiofiles>=23.1.0",
-    "asyncpg>=0.28.0",
 ]
 # Convenience groups
 all = [
-    "ticker-converter[storage,database,airflow,models,api,dev,cloud,viz,auth,monitoring,async]"
+    "ticker-converter[database,api,dev]"
 ]
 
 [project.scripts]
@@ -143,7 +88,7 @@ where = ["src"]
 # Black code formatting
 [tool.black]
 line-length = 88
-target-version = ['py39']
+target-version = ['py311']
 include = '\.pyi?$'
 extend-exclude = '''
 /(
@@ -161,7 +106,7 @@ extend-exclude = '''
 
 # Ruff linting configuration
 [tool.ruff]
-target-version = "py39"
+target-version = "py311"
 line-length = 88
 
 [tool.ruff.lint]
@@ -260,8 +205,6 @@ disable = [
 [tool.pylint.typecheck]
 generated-members = [
     "pydantic.*",
-    "src.ticker_converter.data_models.quality_metrics.DataQualityMetrics.*",
-    "src.ticker_converter.data_models.quality_metrics.DataQualityReport.*",
 ]
 ignored-classes = ["FieldInfo"]
