+++ src/ticker_converter/database/__init__.py
@@ -1,0 +1,1 @@
+"""Database utilities and connection management."""
