+++ src/ticker_converter/etl/__init__.py
@@ -1,0 +1,1 @@
+"""ETL modules for data extraction, transformation, and loading."""
