--- tests/integration/storage/__init__.py
@@ -1,1 +1,0 @@
-"""Integration tests init."""
