+++ tests/unit/api/test_endpoints.py
@@ -1,0 +1,21 @@
+"""Tests for API endpoints."""
+
+
+class TestAPI:
+    """Test API endpoints."""
+    
+    def test_health_check_placeholder(self):
+        """Placeholder test for health check endpoint."""
+        # This test would require setting up the FastAPI app and database
+        # For now, it's a placeholder to maintain test structure
+        assert True
+        
+    def test_top_performers_placeholder(self):
+        """Placeholder test for top performers endpoint."""
+        # This test would verify the top performers endpoint returns correct data
+        assert True
+        
+    def test_price_ranges_placeholder(self):
+        """Placeholder test for price ranges endpoint."""
+        # This test would verify the price ranges endpoint returns correct data
+        assert True
