+++ tests/unit/sql/test_queries.py
@@ -1,0 +1,20 @@
+"""Tests for SQL queries and database operations."""
+
+
+class TestSQLQueries:
+    """Test SQL queries and database operations."""
+    
+    def test_dimension_tables_placeholder(self):
+        """Placeholder test for dimension table queries."""
+        # This test would verify dimension table creation and population
+        assert True
+        
+    def test_fact_tables_placeholder(self):
+        """Placeholder test for fact table queries."""
+        # This test would verify fact table operations
+        assert True
+        
+    def test_analytical_views_placeholder(self):
+        """Placeholder test for analytical views."""
+        # This test would verify analytical views work correctly
+        assert True
